requirements.txt
crewai
streamlit>=1.29.0
openai>=1.6.1
unstructured>=0.11.0
pyowm>=3.3.0
langchain
langchain-community
langchain-core
langchain-groq
tools>=0.1.9
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.10
python-dotenv>=1.0.0
langchain-openai>=0.0.5
sympy